requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.4
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cmake==3.26.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
DLLogger==1.0.0
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
g2p-en==2.1.0
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.1.2
google-auth==2.6.0
google-auth-oauthlib==0.5.2
grpcio==1.48.2
huggingface-hub==0.14.1
hydra-core==1.2.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.2.0
inflect==6.0.4
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
jiwer==3.0.1
joblib==1.1.1
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
librosa==0.10.0.post2
lightning-utilities==0.8.0
lit==16.0.3
llvmlite==0.39.1
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.2
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nemo-toolkit==1.18.0rc0
nest-asyncio==1.5.6
networkx==2.8.4
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.14.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
plac==1.3.5
platformdirs==2.5.2
ply==3.11
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2022.7
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.2.0
rapidfuzz==2.13.7
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.0
rich==13.3.5
rsa==4.7.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.22.2
setproctitle==1.3.2
setuptools==66.0.0
shellingham==1.5.0.post1
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
sox==1.4.1
soxr==0.3.5
stack-data==0.2.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
texterrors==0.4.4
threadpoolctl==2.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.0
torchmetrics==0.11.4
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.29.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.2
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wget==3.2
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.1
youtokentome==1.0.6
zipp==3.11.0
